src/python-ai-service/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.0
pydantic-settings==2.1.0
openai>=1.0
python-dotenv==1.0.0
requests==2.31.0
httpx>=0.27.0,<1.0.0
